+++ CompositeDulcesTest/ProductoTest.cs
@@ -7,7 +7,7 @@ namespace DomainTest
     public class ProductoTest
     {        
         [TestCaseSource("DatosIncorrectosParaProducto")]
-        public void PuedeCrearProducto(double cantidad, double costoUnitario,double porcentajeDeUtilidad,
+        public void PuedeCrearProducto(decimal cantidad, decimal costoUnitario,decimal porcentajeDeUtilidad,
             string esperado)
         {
             var obtenido = ProductoPuedeCrear.PuedeCrearProducto(cantidad, costoUnitario, porcentajeDeUtilidad);
@@ -15,15 +15,15 @@ namespace DomainTest
         }
         private static IEnumerable<TestCaseData> DatosIncorrectosParaProducto()
         {
-            yield return new TestCaseData(-5, -560,0,
+            yield return new TestCaseData(-5m, -560m,0m,
                 "Cantidad invalida, Costo unitario invalido").SetName("CrearProductoCantidadYCostoInvalido");
-            yield return new TestCaseData(5, -560,0,
+            yield return new TestCaseData(5m, -560m,0m,
                 "Costo unitario invalido").SetName("CrearProductoCostoInvalido");
-            yield return new TestCaseData(5, 560,1000,
+            yield return new TestCaseData(5m, 560m,1000m,
                 "Porcentaje de utilidad invalido").SetName("CrearProductoPorcentajeDeUtilidadMayorQueCien");
-            yield return new TestCaseData(5, 560,-1000,
+            yield return new TestCaseData(5m, 560m,-1000m,
                 "Porcentaje de utilidad invalido").SetName("CrearProductoPorcentajeDeUtilidadNegativo");
-            yield return new TestCaseData(-5, 560,0,
+            yield return new TestCaseData(-5m, 560m,0m,
                 "Cantidad invalida").SetName("CrearProductoCantidadInvalida");
         }
     }
